requirements-dev.txt
-r requirements.txt
pylint==3.3.4
pytest==7.4.3
pytest-django==4.7.0
